Extract lists into a standalone test
