+++ crates/test-rust-wasm/src/bin/lists.rs
@@ -1,0 +1,3 @@
+include!("../../../../tests/runtime/lists/wasm.rs");
+
+fn main() {}
