+++ tests/runtime/exports.c
@@ -5,118 +5,6 @@
 #include <string.h>
 #include <wasm.h>
 
-// "custom allocator" which just keeps track of allocated bytes
-
-static size_t ALLOCATED_BYTES = 0;
-
-__attribute__((export_name("canonical_abi_realloc")))
-void *canonical_abi_realloc( void *ptr, size_t orig_size, size_t orig_align, size_t new_size) {
-  void *ret = realloc(ptr, new_size);
-  if (!ret)
-    abort();
-  ALLOCATED_BYTES -= orig_size;
-  ALLOCATED_BYTES += new_size;
-  return ret;
-}
-
-__attribute__((export_name("canonical_abi_free")))
-void canonical_abi_free(void *ptr, size_t size, size_t align) {
-  if (size > 0) {
-    ALLOCATED_BYTES -= size;
-    free(ptr);
-  }
-}
-
-uint32_t wasm_allocated_bytes(void) {
-  return ALLOCATED_BYTES;
-}
-
-
-void wasm_list_param(wasm_list_u8_t *a) {
-  assert(a->len == 4);
-  assert(a->ptr[0] == 1);
-  assert(a->ptr[1] == 2);
-  assert(a->ptr[2] == 3);
-  assert(a->ptr[3] == 4);
-  wasm_list_u8_free(a);
-}
-
-void wasm_list_param2(wasm_string_t *a) {
-  assert(a->len == 3);
-  assert(a->ptr[0] == 'f');
-  assert(a->ptr[1] == 'o');
-  assert(a->ptr[2] == 'o');
-  wasm_string_free(a);
-}
-
-void wasm_list_param3(wasm_list_string_t *a) {
-  assert(a->len == 3);
-  assert(a->ptr[0].len == 3);
-  assert(a->ptr[0].ptr[0] == 'f');
-  assert(a->ptr[0].ptr[1] == 'o');
-  assert(a->ptr[0].ptr[2] == 'o');
-
-  assert(a->ptr[1].len == 3);
-  assert(a->ptr[1].ptr[0] == 'b');
-  assert(a->ptr[1].ptr[1] == 'a');
-  assert(a->ptr[1].ptr[2] == 'r');
-
-  assert(a->ptr[2].len == 3);
-  assert(a->ptr[2].ptr[0] == 'b');
-  assert(a->ptr[2].ptr[1] == 'a');
-  assert(a->ptr[2].ptr[2] == 'z');
-
-  wasm_list_string_free(a);
-}
-
-void wasm_list_param4(wasm_list_list_string_t *a) {
-  assert(a->len == 2);
-  assert(a->ptr[0].len == 2);
-  assert(a->ptr[1].len == 1);
-
-  assert(a->ptr[0].ptr[0].len == 3);
-  assert(a->ptr[0].ptr[0].ptr[0] == 'f');
-  assert(a->ptr[0].ptr[0].ptr[1] == 'o');
-  assert(a->ptr[0].ptr[0].ptr[2] == 'o');
-
-  assert(a->ptr[0].ptr[1].len == 3);
-  assert(a->ptr[0].ptr[1].ptr[0] == 'b');
-  assert(a->ptr[0].ptr[1].ptr[1] == 'a');
-  assert(a->ptr[0].ptr[1].ptr[2] == 'r');
-
-  assert(a->ptr[1].ptr[0].len == 3);
-  assert(a->ptr[1].ptr[0].ptr[0] == 'b');
-  assert(a->ptr[1].ptr[0].ptr[1] == 'a');
-  assert(a->ptr[1].ptr[0].ptr[2] == 'z');
-
-  wasm_list_list_string_free(a);
-}
-
-void wasm_list_result(wasm_list_u8_t *ret0) {
-  ret0->ptr = canonical_abi_realloc(NULL, 0, 1, 5);
-  ret0->len = 5;
-  ret0->ptr[0] = 1;
-  ret0->ptr[1] = 2;
-  ret0->ptr[2] = 3;
-  ret0->ptr[3] = 4;
-  ret0->ptr[4] = 5;
-}
-
-void wasm_list_result2(wasm_string_t *ret0) {
-  wasm_string_dup(ret0, "hello!");
-}
-
-void wasm_list_result3(wasm_list_string_t *ret0) {
-  ret0->len = 2;
-  ret0->ptr = canonical_abi_realloc(NULL, 0, alignof(wasm_string_t), 2 * sizeof(wasm_string_t));
-
-  wasm_string_dup(&ret0->ptr[0], "hello,");
-  wasm_string_dup(&ret0->ptr[1], "world!");
-}
-
-void wasm_string_roundtrip(wasm_string_t *a, wasm_string_t *ret0) {
-  *ret0 = *a;
-}
 
 wasm_wasm_state_t wasm_wasm_state_create(void) {
   return wasm_wasm_state_new((void*) 100);
